[修正]元素拖曳到 pdf 時的錯誤
