--- src/components/drag-box/drag-box.style.js
@@ -1,19 +1,0 @@
-import styled from "styled-components";
-
-export const StyledDragBox = styled.div`
-  position: relative;
-  background-color: burlywood;
-  padding: 10px;
-`;
-
-export const DragItem = styled.div`
-  width: 100px;
-  padding: 5px 0;
-  margin: 0.5rem;
-  justify-content: center;
-  background-color: orange;
-  color: black;
-  z-index: 1;
-  text-align: center;
-  cursor: cursor;
-`;
