+++ src/components/pdf-pages/pdf-pages.component.jsx
@@ -1,12 +1,14 @@
 import { useState, useEffect } from "react";
+import { useSelector } from "react-redux";
 
 import PdfPage from "@/components/pdf-page/pdf-page.component";
+
+import { selectPdfId } from "@/store/pdf/selector";
 
 const PdfPages = (props) => {
   const { getPdfPage, pageCounts, canvasRefs, scale } = props;
   const [pages, setPages] = useState([]);
-
-  console.log("pageCounts", pageCounts);
+  const pdfId = useSelector(selectPdfId);
 
   useEffect(() => {
     if (!pageCounts) return;
@@ -22,7 +24,7 @@ const PdfPages = (props) => {
       {pages.map((page, index) => (
         <PdfPage
           key={index + 1}
-          dropId={`page-${index + 1}`}
+          dropId={`${pdfId}-${index + 1}`}
           asyncPage={page}
           scale={scale}
         />
