--- src/components/viewer/viewer.style.js
@@ -1,5 +1,0 @@
-import styled from "styled-components";
-
-export const Container = styled.div`
-  display: flex;
-`;
