+++ src/store/pdf/selector.js
@@ -1,3 +1,5 @@
 export const selectPdf = (state) => state.pdf.pdf;
 
 export const selectNumPages = (state) => state.pdf.numPages;
+
+export const selectPdfId = (state) => state.pdf.id;
